+++ src/modules/content/component/placeholder.component.tsx
@@ -1,0 +1,7 @@
+import { Result } from 'antd';
+
+export default function PlaceholderComponent() {
+  return (
+    <Result status="info" title="Select a content in the explorer"></Result>
+  );
+}