+++ src/modules/content/interface/content-module.interface.ts
@@ -1,0 +1,5 @@
+export interface IContentModule {
+  id: string;
+  name: string;
+  tags: string[];
+}